requirements.txt
altgraph==0.17.4
asgiref==3.8.1
bidict==0.23.1
blinker==1.8.2
cffi==1.17.1
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Django==5.1.6
django-otp==1.5.4
dnspython==2.7.0
eventlet==0.38.2
Flask==3.1.0
greenlet==3.1.1
h11==0.14.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
packaging==24.2
pefile==2023.2.7
pycparser==2.22
pycryptodome==3.21.0
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.11
pyotp==2.9.0
python-dotenv==1.1.0
python-engineio==4.11.1
python-socketio==5.12.0
pywin32-ctypes==0.2.3
scapy==2.6.1
setuptools==75.8.0
simple-websocket==1.1.0
sqlparse==0.5.1
tzdata==2024.1
Werkzeug==3.0.3
wsproto==1.2.0
